@@ -10,6 +10,7 @@ open Microsoft.AspNetCore.HttpsPolicy;
 open Microsoft.AspNetCore.Mvc
 open Microsoft.Extensions.Configuration
 open Microsoft.Extensions.DependencyInjection
+open Microsoft.Extensions.Hosting
 
 type Startup private () =
     new (configuration: IConfiguration) as this =
@@ -24,10 +25,11 @@ type Startup private () =
         ) |> ignore
 
         // Add framework services.
-        services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2) |> ignore
+        services.AddControllersWithViews().AddRazorRuntimeCompilation() |> ignore
+        services.AddRazorPages() |> ignore
 
     // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-    member this.Configure(app: IApplicationBuilder, env: IHostingEnvironment) =
+    member this.Configure(app: IApplicationBuilder, env: IWebHostEnvironment) =
 
         if (env.IsDevelopment()) then
             app.UseDeveloperExceptionPage() |> ignore
@@ -37,10 +39,12 @@ type Startup private () =
         app.UseDefaultFiles() |> ignore
         app.UseStaticFiles() |> ignore
 
-        app.UseMvc(fun routes ->
-            routes.MapRoute(
+        app.UseRouting() |> ignore
+
+        app.UseEndpoints(fun endpoints ->
+            endpoints.MapControllerRoute(
                 name = "default",
-                template = "{controller=Home}/{action=Index}/{id?}") |> ignore
-            ) |> ignore
+                pattern = "{controller=Home}/{action=Index}/{id?}") |> ignore
+            endpoints.MapRazorPages() |> ignore) |> ignore
 
     member val Configuration : IConfiguration = null with get, set
